refactor audio modules into package
