+++ OcchioOnniveggente/src/main.py
@@ -23,7 +23,8 @@ import logging
 
 from src.config import Settings, get_openai_api_key
 from src.filters import ProfanityFilter
-from src.audio import AudioPreprocessor, record_until_silence, play_and_pulse
+from src.audio.processing import AudioPreprocessor
+from src.audio.recording import record_until_silence, play_and_pulse
 from src.lights import SacnLight, WledLight, color_from_text
 from src.oracle import (
     transcribe,
@@ -35,14 +36,13 @@ from src.oracle import (
     detect_language,
 )
 from src.domain import validate_question
-from src.hotword import is_wake, matches_hotword_text, strip_hotword_prefix
-from src.hotword import is_wake
+from src.audio.hotword import is_wake, matches_hotword_text, strip_hotword_prefix
 from src.chat import ChatState
 from src.conversation import ConversationManager, DialogState
 from src.logging_utils import setup_logging
 from src.language_session import update_language
 from src.cli import _ensure_utf8_stdout, say, oracle_greeting, default_response
-from src.audio_device import pick_device, debug_print_devices
+from src.audio.audio_device import pick_device, debug_print_devices
 from src.profile_utils import get_active_profile, make_domain_settings
 
 logger = logging.getLogger(__name__)
