+++ pyproject.toml
@@ -10,10 +10,10 @@ readme = "README.md"
 requires-python = ">=3.10"
 
 [project.entry-points."oracolo.stt"]
-local = "OcchioOnniveggente.src.audio:create_local_stt"
+local = "OcchioOnniveggente.src.audio.protocols:create_local_stt"
 
 [project.entry-points."oracolo.tts"]
-local = "OcchioOnniveggente.src.audio:create_local_tts"
+local = "OcchioOnniveggente.src.audio.protocols:create_local_tts"
 
 [project.entry-points."oracolo.llm"]
 openai = "OcchioOnniveggente.src.openai_async:create_openai_llm"
